requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
aiogram==3.13.0
python-dotenv==1.0.1